// frontend/src/lib/team-logo-registry.generated.ts
export type EspnLogoLeague =
  | 'nba'
  | 'wnba'
  | 'nhl'
  | 'epl'
  | 'mls'
  | 'ucl'
  | 'ncaaf'
  | 'ncaab';